+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Race Protocol: A multi-chain infrastructure for asymmetric competitive games
 
+# Master(Unreleased)
+
+- Add local blacklist for malformed game address.  Add `disable_blacklist = true` to `[transactor]` section in config file to disable.
+
 # 0.1.0
 
 It basically works with a Holdem game